remove duplicated version info
* rename `__init__.py` to `main.py`
