+++ wafw00f/bin/wafw00f
@@ -5,4 +5,4 @@ from wafw00f import main
 
 
 if __name__ == '__main__':
-    main()
+    main.main()
